+++ packages/mobile/CHANGELOG.md
@@ -8,6 +8,30 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.33.0 (12/18/2025 PST)
+
+#### 🚀 Updates
+
+- New digitTransitionVariant for RollingNumber. [[#237](https://github.com/coinbase/cds/pull/237)]
+
+## 8.32.3 (12/18/2025 PST)
+
+#### 🐞 Fixes
+
+- Fix: update package exports. [[#261](https://github.com/coinbase/cds/pull/261)]
+
+## 8.32.2 (12/18/2025 PST)
+
+#### 🐞 Fixes
+
+- Add fuse.js dependency. [[#256](https://github.com/coinbase/cds/pull/256)]
+
+## 8.32.1 (12/17/2025 PST)
+
+#### 🐞 Fixes
+
+- Fix: update react-native export pointer. [[#259](https://github.com/coinbase/cds/pull/259)]
+
 ## 8.32.0 ((12/16/2025, 08:20 PM PST))
 
 This is an artificial version bump with no new change.